Move the request body into its own class
There are two GObject classes now: BasePayloadPane is the top level
abstract class and RawPayloadPane is the implementation.

BasePayloadPane defines a property called `payload`. Each implementation
shall define its own getter and setter to allow other widgets to take or
set the contents of the pane, but each implementation is free to do so
on its own way. The RawPayloadPane just takes and sets the SourceView
buffer contents, but other implementations may parse the byte array
differently.